feat: simplifies even further
