+++ boilerplate/frontend/react/src/App.tsx
@@ -7,42 +7,42 @@ import Dashboard from "./Dashboard";
 import { PreBuiltUIList, SuperTokensConfig, ComponentWrapper } from "./config";
 import Home from "./Home";
 
-// Initialize SuperTokens - ideally in the global
+// Initialize SuperTokens - ideally in the global scope
 SuperTokens.init(SuperTokensConfig);
 
 function App() {
     return (
         <SuperTokensWrapper>
-            <header>
-                <nav className="header-container">
-                    <Link to="/">
-                        <img src="/ST.svg" alt="SuperTokens" />
-                    </Link>
-                    <ul className="header-container-right">
-                        <li>
-                            <a
-                                href="https://supertokens.com/docs/guides/getting-started/react"
-                                target="_blank"
-                                rel="noopener noreferrer"
-                            >
-                                Docs
-                            </a>
-                        </li>
-                        <li>
-                            <a
-                                href="https://github.com/supertokens/create-supertokens-app"
-                                target="_blank"
-                                rel="noopener noreferrer"
-                            >
-                                CLI Repo
-                            </a>
-                        </li>
-                    </ul>
-                </nav>
-            </header>
-            <ComponentWrapper>
+            <BrowserRouter>
                 <main className="App app-container">
-                    <BrowserRouter>
+                    <header>
+                        <nav className="header-container">
+                            <Link to="/">
+                                <img src="/ST.svg" alt="SuperTokens" />
+                            </Link>
+                            <ul className="header-container-right">
+                                <li>
+                                    <a
+                                        href="https://supertokens.com/docs/guides/getting-started/react"
+                                        target="_blank"
+                                        rel="noopener noreferrer"
+                                    >
+                                        Docs
+                                    </a>
+                                </li>
+                                <li>
+                                    <a
+                                        href="https://github.com/supertokens/create-supertokens-app"
+                                        target="_blank"
+                                        rel="noopener noreferrer"
+                                    >
+                                        CLI Repo
+                                    </a>
+                                </li>
+                            </ul>
+                        </nav>
+                    </header>
+                    <ComponentWrapper>
                         <Routes>
                             <Route path="/" element={<Home />} />
                             {/* This shows the login UI on "/auth" route */}
@@ -60,9 +60,9 @@ function App() {
                                 }
                             />
                         </Routes>
-                    </BrowserRouter>
+                    </ComponentWrapper>
                 </main>
-            </ComponentWrapper>
+            </BrowserRouter>
         </SuperTokensWrapper>
     );
 }
