--- boilerplate/frontend/react/src/Dashboard/Footer.tsx
@@ -1,24 +1,0 @@
-import { SeparatorLine } from "../assets/images";
-import { Link } from "./index";
-
-interface FooterProps {
-    links: Link[];
-}
-
-export default function Footer({ links }: FooterProps) {
-    return (
-        <>
-            <div className="bottom-links-container">
-                {links.map((link) => (
-                    <div className="link" key={link.name}>
-                        <img className="link-icon" src={link.icon} alt={link.name} />
-                        <div role={"button"} onClick={link.onClick}>
-                            {link.name}
-                        </div>
-                    </div>
-                ))}
-            </div>
-            <img className="separator-line" src={SeparatorLine} alt="separator" />
-        </>
-    );
-}
