+++ boilerplate/frontend/react/src/Dashboard/index.tsx
@@ -1,16 +1,6 @@
-import { useSessionContext } from "supertokens-auth-react/recipe/session";
-import { signOut } from "supertokens-auth-react/recipe/session";
-import { recipeDetails } from "../config";
-import SessionInfo from "./SessionInfo";
-import { BlogsIcon, CelebrateIcon, GuideIcon, SignOutIcon } from "../assets/images";
+import { useSessionContext, signOut } from "supertokens-auth-react/recipe/session";
 import { useNavigate } from "react-router-dom";
-import Footer from "./Footer";
-
-export interface Link {
-    name: string;
-    onClick: () => void;
-    icon: string;
-}
+import { getApiDomain } from "../config";
 
 export default function Dashboard() {
     const sessionContext = useSessionContext();
@@ -25,43 +15,53 @@ export default function Dashboard() {
         navigate("/auth");
     }
 
-    function openLink(url: string) {
-        window.open(url, "_blank");
+    async function callAPIClicked() {
+        try {
+            const response = await fetch(getApiDomain() + "/sessioninfo");
+            const data = await response.json();
+            window.alert("Session Information:\n" + JSON.stringify(data, null, 2));
+        } catch (err) {
+            window.alert("Error calling API: " + err.message);
+        }
     }
-
-    const links: Link[] = [
-        {
-            name: "Blogs",
-            onClick: () => openLink("https://supertokens.com/blog"),
-            icon: BlogsIcon,
-        },
-        {
-            name: "Documentation",
-            onClick: () => openLink(recipeDetails.docsLink),
-            icon: GuideIcon,
-        },
-        {
-            name: "Sign Out",
-            onClick: logoutClicked,
-            icon: SignOutIcon,
-        },
-    ];
 
     return (
         <div className="fill" id="home-container">
             <div className="main-container">
                 <div className="top-band success-title bold-500">
-                    <img src={CelebrateIcon} alt="Login successful" className="success-icon" /> Login successful
+                    <img src="/celebrate-icon.svg" alt="Login successful" className="success-icon" />
+                    Login successful
                 </div>
                 <div className="inner-content">
                     <div>Your userID is:</div>
                     <div className="truncate" id="user-id">
                         {sessionContext.userId}
                     </div>
-                    <SessionInfo />
+                    <button onClick={callAPIClicked} className="sessionButton">
+                        Call API
+                    </button>
                 </div>
             </div>
-            <Footer links={links} />
+            <nav className="bottom-links-container">
+                <a href="https://supertokens.com/blog" target="_blank" rel="noopener noreferrer" className="link">
+                    <img src="/blogs-icon.svg" alt="Blogs" className="link-icon" />
+                    <span>Blogs</span>
+                </a>
+                <a
+                    href="https://supertokens.com/docs/guides"
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className="link"
+                >
+                    <img src="/guide-icon.svg" alt="Documentation" className="link-icon" />
+                    <span>Documentation</span>
+                </a>
+                <button onClick={logoutClicked} className="link">
+                    <img src="/signout-icon.svg" alt="Sign Out" className="link-icon" />
+                    <span>Sign Out</span>
+                </button>
+            </nav>
+            <img className="separator-line" src="/separator-line.svg" alt="separator" />
         </div>
     );
 }
